+++ packages/canvas/src-native/canvas-android/canvas/src/main/java/org/nativescript/canvas/NSCSVG.kt
@@ -52,8 +52,8 @@ class NSCSVG : View {
 		if (sync) {
 			NSCCanvas.nativeResizeCustomSurface(
 				svgCanvas,
-				w.toFloat(),
-				h.toFloat(),
+				w,
+				h,
 				metrics.density,
 				true,
 				metrics.densityDpi
@@ -78,8 +78,8 @@ class NSCSVG : View {
 		currentTask = executor.submit {
 			NSCCanvas.nativeResizeCustomSurface(
 				svgCanvas,
-				w.toFloat(),
-				h.toFloat(),
+				w,
+				h,
 				metrics.density,
 				true,
 				metrics.densityDpi
@@ -116,8 +116,8 @@ class NSCSVG : View {
 				if (sync) {
 
 					svgCanvas = NSCCanvas.nativeInitContextWithCustomSurface(
-						w.toFloat(),
-						h.toFloat(),
+						w,
+						h,
 						metrics.density,
 						true,
 						Color.BLACK,
@@ -148,8 +148,8 @@ class NSCSVG : View {
 				currentTask = executor.submit {
 					synchronized(lock) {
 						svgCanvas = NSCCanvas.nativeInitContextWithCustomSurface(
-							w.toFloat(),
-							h.toFloat(),
+							w,
+							h,
 							metrics.density,
 							true,
 							Color.BLACK,
